autoscaling-app/src/requirements.txt
Flask==3.0.0
Pillow==10.1.0
boto3==1.34.0
psutil==5.9.6